Fix datamodel gate
Fixes #41
