+++ .github/workflows/checks.yml
@@ -55,8 +55,8 @@ jobs:
       - name: Privacy gate - no absolute user paths in tracked files
         run: uv run python scripts/set_data_folder.py --check
 
-      - name: M syntax gate - catch Desktop's unlocalised "Token ',' expected" first
-        run: uv run python scripts/check_m_syntax.py --all
+      - name: Data model gate - catch M and TMDL structural defects first
+        run: uv run python scripts/check_datamodel.py --all
 
       - name: Shared conventions reach every agent (a subagent sees ONLY its persona)
         run: uv run python scripts/sync_agent_conventions.py --check
